Move test to separate dir, use canonical output
