+++ cloud/filestore/tests/loadtest/max-file-count-test/canondata/result.json
@@ -1,0 +1,5 @@
+{
+    "test.test_max_file_count": {
+        "uri": "file://test.test_max_file_count/results.txt"
+    }
+}
