+++ cloud/filestore/tests/loadtest/max-file-count-test/canondata/test.test_max_file_count/results.txt
@@ -1,0 +1,3 @@
+/	max-file-count-loadtest:<random_guid>
+/	max-file-count-loadtest:<random_guid>
+/	max-file-count-loadtest:<random_guid>
