+++ cloud/filestore/tests/loadtest/max-file-count-test/max-file-count.txt
@@ -1,6 +1,7 @@
 Tests {
     LoadTest {
         Name: "max-file-count"
+        ClientId: "max-file-count-loadtest"
         CreateFileStoreRequest {
             FileSystemId: "max-file-count"
             FolderId: "folder"
