+++ cloud/filestore/tests/loadtest/max-file-count-test/ya.make
@@ -1,0 +1,25 @@
+PY3TEST()
+
+INCLUDE(${ARCADIA_ROOT}/cloud/filestore/tests/recipes/medium.inc)
+
+TEST_SRCS(
+    test.py
+)
+
+DEPENDS(
+    cloud/filestore/apps/client
+    cloud/filestore/tools/testing/loadtest/bin
+)
+
+PEERDIR(
+    cloud/filestore/tests/python/lib
+)
+
+SET(
+    NFS_STORAGE_CONFIG_PATCH
+    cloud/filestore/tests/common_configs/nfs-storage-migration-patch.txt
+)
+
+INCLUDE(${ARCADIA_ROOT}/cloud/filestore/tests/recipes/service-kikimr.inc)
+
+END()
